// src/common/enums.ts
export enum typeAuth {
    LOGIN = 'login',
    LOGOUT = 'logout',
    REGISTER = 'register',
    UPDATE = 'update',
    DELETE = 'delete',
}

export enum methodAuth {
    TOKEN= 'token',
    PASSWORD=  'password',
    
}

export enum statusAuth {
    SUCCESS = 'success',
    ERROR = 'error',
}

export enum userRol {
    MODERADOR = 'moderator',
    USER = 'user',
} 